+++ test/helpers_test.dart
@@ -1,7 +1,12 @@
-import 'package:dartle/src/_log.dart';
+import 'dart:io';
+
+import 'package:dartle/dartle.dart';
 import 'package:dartle/src/_utils.dart';
 import 'package:logging/logging.dart';
+import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
+
+import 'test_utils.dart';
 
 void main([List<String> args = const []]) {
   if (args.contains('log')) {
@@ -105,5 +110,101 @@ void main([List<String> args = const []]) {
         findMatchingByWords(
             'f', const ['fooBar', 'fooBarBaz', 'test', 'check', 'fooBarFoo']),
         isNull);
+  });
+
+  group('deleteAll', () {
+    DartleTestFileSystem fs = createTempFileSystem();
+    Directory? foo, bar;
+    File? txtFile, mdFile, barTxtFile, barMdFile;
+    setUp(() async {
+      fs = createTempFileSystem();
+      foo = fs.directory('foo');
+      txtFile = fs.file(p.join('foo', 'file.txt'));
+      mdFile = fs.file(p.join('foo', 'file.md'));
+      bar = fs.directory(p.join('foo', 'bar'));
+      barTxtFile = fs.file(p.join('foo', 'bar', 'file.txt'));
+      barMdFile = fs.file(p.join('foo', 'bar', 'file.md'));
+      await bar!.create(recursive: true);
+      for (var f in [txtFile!, mdFile!, barTxtFile!, barMdFile!]) {
+        await f.create();
+      }
+    });
+
+    test('full directory', () async {
+      // make sure the file tree was created correctly
+      expect(
+          await Directory(fs.root)
+              .list(recursive: true)
+              .map((e) => e.path)
+              .toSet(),
+          equals({foo!, bar!, txtFile!, mdFile!, barTxtFile!, barMdFile!}
+              .map((e) => e.path)
+              .toSet()));
+
+      // actual test
+      await withFileSystem(fs, () async {
+        await deleteAll(dir('foo'));
+      });
+      expect(await Directory(fs.root).list(recursive: true).toList(), isEmpty);
+    });
+
+    test('sub-directory', () async {
+      await withFileSystem(fs, () async {
+        await deleteAll(dir(p.join('foo', 'bar')));
+      });
+      expect(
+          await Directory(fs.root)
+              .list(recursive: true)
+              .map((e) => e.path)
+              .toSet(),
+          equals({foo!, txtFile!, mdFile!}.map((e) => e.path).toSet()));
+    });
+
+    test('directory with pattern', () async {
+      await withFileSystem(fs, () async {
+        await deleteAll(dir('foo', fileExtensions: {'.md', '.jpg'}));
+      });
+      expect(
+          await Directory(fs.root)
+              .list(recursive: true)
+              .map((e) => e.path)
+              .toSet(),
+          equals(
+              {foo!, bar!, txtFile!, barTxtFile!}.map((e) => e.path).toSet()));
+    });
+
+    test('directory with pattern matching everything', () async {
+      await withFileSystem(fs, () async {
+        await deleteAll(dir('foo', fileExtensions: {'.md', '.txt'}));
+      });
+      expect(await Directory(fs.root).list(recursive: true).toSet(), isEmpty);
+    });
+
+    test('directory with file exclusion', () async {
+      await withFileSystem(fs, () async {
+        await deleteAll(dir('foo', exclusions: const {'file.txt'}));
+      });
+      expect(
+          await Directory(fs.root)
+              .list(recursive: true)
+              .map((e) => e.path)
+              .toSet(),
+          equals(
+              {foo!, bar!, txtFile!, barTxtFile!}.map((e) => e.path).toSet()));
+    });
+
+    test('directory with directory exclusion', () async {
+      await withFileSystem(fs, () async {
+        await deleteAll(dir('foo', exclusions: const {'bar'}));
+      });
+      expect(
+          await Directory(fs.root)
+              .list(recursive: true)
+              .map((e) => e.path)
+              .toSet(),
+          equals({foo!, bar!, barTxtFile!, barMdFile!}
+              .map((e) => e.path)
+              .toSet()));
+    });
   });
 }